fix:removing pages router and fixing styles
