+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "@fullcalendar/core": "^6.1.15",
     "@fullcalendar/daygrid": "^6.1.15",
-    "@fullcalendar/interaction": "^6.1.14",
+    "@fullcalendar/interaction": "^6.1.15",
     "@fullcalendar/react": "^6.1.14",
     "@fullcalendar/resource": "^6.1.14",
     "@fullcalendar/resource-timeline": "^6.1.14",
