+++ src/app/(auth)/login/page.tsx
@@ -1,0 +1,13 @@
+import LoginForm from "@/components/auth/login-form";
+import LoginComp from "@/components/auth/login-form";
+
+// SignIn component
+const SignIn = () => {
+  return (
+    <div className="flex flex-col justify-center text-gray-800 items-center h-full">
+      <LoginForm />
+    </div>
+  );
+};
+
+export default SignIn;
