+++ src/app/(auth)/signUp/page.tsx
@@ -1,0 +1,11 @@
+import SignupForm from "@/components/auth/signup-form";
+
+const SignUpPage = () => {
+  return (
+    <div className="flex flex-col justify-center text-gray-800 items-center h-full">
+      <SignupForm />
+    </div>
+  );
+};
+
+export default SignUpPage;
