+++ src/app/(dashboard)/calendar/page.tsx
@@ -8,14 +8,14 @@ import interactionPlugin from "@fullcalendar/interaction";
 import timeGridPlugin from "@fullcalendar/timegrid";
 import rrulePlugin from "@fullcalendar/rrule";
 import { DateSelectArg, EventApi, EventContentArg } from "@fullcalendar/core";
-import EventFormDialog from "./EventFormModal";
+import EventFormDialog from "../../../comp/EventFormModal";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import Availability from "../pages/tabs/availability";
-import CreateBookings from "./tabs/create_bookings";
-import { auth, db } from "../../firebase";
-import { createEvent } from "../services/userService";
-import useFetchEvents from "../hooks/useFetchEvents";
-import { EventInput } from "../interfaces/types";
+import Availability from "../../../comp/tabs/availability";
+import CreateBookings from "../../../comp/tabs/create_bookings";
+import { auth, db } from "../../../../firebase";
+import { createEvent } from "../../../services/userService";
+import useFetchEvents from "../../../hooks/useFetchEvents";
+import { EventInput } from "../../../interfaces/types";
 import { addDoc, getDoc, Timestamp } from "firebase/firestore";
 import { doc, updateDoc } from "firebase/firestore";
 import { collection, query, getDocs } from "firebase/firestore";
