+++ src/app/(dashboard)/layout.tsx
@@ -1,0 +1,37 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import { Montserrat } from "next/font/google";
+import { Source_Sans_3 } from "next/font/google";
+import { Exo_2 } from "next/font/google";
+import "../globals.css";
+
+import Header from "@/components/header";
+import Footer from "@/components/footer";
+
+const exo_2 = Exo_2({
+  subsets: ["latin"],
+  weight: "700",
+  style: "normal",
+  display: "swap",
+});
+
+export const metadata: Metadata = {
+  title: "Rebus Pro - Dashboard",
+  description: "Rebus Pro - The best way to manage your events",
+};
+
+export default function DashboardLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body className={`${exo_2.className} min-h-screen flex flex-col`}>
+        <Header />
+        <main className="flex-grow">{children}</main>
+        <Footer />
+      </body>
+    </html>
+  );
+}
