--- src/pages/_app.tsx
@@ -1,30 +1,0 @@
-import "../app/globals.css"; // Adjust the import path as needed
-import Layout from "../app/layout"; // Adjust the import path as needed
-
-import Header from "@/components/header";
-import Footer from "@/components/footer";
-import { Toaster } from "@/components/ui/toaster";
-import { SpeedInsights } from "@vercel/speed-insights/next";
-
-function MyApp({ Component, pageProps }: { Component: any; pageProps: any }) {
-  return (
-    // <>
-    //   <Header />
-    //   <main>
-    //     <Component {...pageProps} />
-    //   </main>
-    //   <Toaster />
-    //   <Footer />
-    // </>
-    <div className="min-h-screen flex flex-col">
-      <Header />
-      <main className="flex-grow">
-        <Component {...pageProps} />
-        <SpeedInsights />
-      </main>
-      <Footer />
-    </div>
-  );
-}
-
-export default MyApp;
