--- src/pages/home.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const home = () => {
-  return (
-    <div>
-      {/* Your home page content goes here */}
-      <h1>Welcome to the Home Page!</h1>
-    </div>
-  );
-};
-
-export default home;
